% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_image.R
\name{to_image}
\alias{to_image}
\title{Add image path to a data frame}
\usage{
to_image(data, column_name = "state", dir = "states")
}
\arguments{
\item{data}{your data frame}

\item{column_name}{the column containing values you want to convert to an image path}
}
\description{
This function allows you to add an image path to a data frame based on the data contained in a column of the data frame
}
\examples{
to_image <- function(data = d, column_name = "state")
}
\keyword{image}
