% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_state_flag_tree.R
\name{us_state_flag_tree}
\alias{us_state_flag_tree}
\title{Annotate phylogenetic trees with USA state flags}
\usage{
us_state_flag_tree(
  tree,
  metadata,
  image_column = "image_path",
  label_column = "label"
)
}
\arguments{
\item{tree}{tree object}

\item{metadata}{metadata associated with tree}

\item{image_column}{column containing pathway to state flag icons}

\item{label_column}{column containing tip labels (ex. taxonomic name, or accession number)}
}
\description{
This function allows you to plot a phylogenetic tree annotated with the United States of America state flag that corresponds with a tip.
}
\examples{
us_state_flag_tree(tree = t, metadata = d, image_column = "image_path", label_column = "label")
}
\keyword{tree}
