% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo101-package.R
\docType{package}
\name{phylo101-package}
\alias{phylo101}
\alias{phylo101-package}
\title{phylo101: Introduction to Phylogenetics}
\description{
This package provides an introduction to phylogenetic analysis.
}
\author{
\strong{Maintainer}: John Hinkle \email{john.hinkle@utexas.edu}

Authors:
\itemize{
  \item Madina Seidualy \email{madise@utexas.edu}
  \item Monica Dooley \email{monica-dooley@utexas.edu}
}

}
\keyword{internal}
