% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo101-package.R
\docType{package}
\name{phylo101-package}
\alias{phylo101}
\alias{phylo101-package}
\title{phylo101: Introduction to Phylogenetics}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: John Hinkle \email{john.hinkle@utexas.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
